requirements_gpu_cuda118.txt
certifi==2022.12.7
charset-normalizer==2.1.1
coloredlogs==15.0.1
filelock==3.9.0
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
onnx==1.14.1
onnxruntime==1.15.1
packaging==23.1
Pillow==9.3.0
protobuf==4.24.2
pyreadline3==3.4.1
requests==2.28.1
sympy==1.11.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
typing_extensions==4.4.0
urllib3==1.26.13
